refactor: api fetch logic
