+++ actions/uploadThing.ts
@@ -1,19 +1,23 @@
-// uploadThing.ts
-export async function uploadFiles(formData: FormData) {
-  try {
-    const response = await fetch('https://experiments-testing.azurewebsites.net/upload', {
-      method: 'POST',
-      body: formData,
-    });
+"use server";
 
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
+import { UTApi } from "uploadthing/server";
 
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    console.error(error);
-    return { error: "Internal server error" };
-  }
+const utapi = new UTApi();
+
+/**
+ * @see https://docs.uploadthing.com/api-reference/server#uploadfiles
+ */
+export async function uploadFiles(fd: FormData) {
+  const files = fd.getAll("file") as File[];
+  const uploadedFiles = await utapi.uploadFiles(files);
+  return uploadedFiles;
+}
+
+/**
+ * @see https://docs.uploadthing.com/api-reference/server#uploadfilesfromurl
+ */
+export async function uploadFromUrl(fd: FormData) {
+  const url = fd.get("url") as string;
+  const uploadedFile = await utapi.uploadFilesFromUrl(url);
+  return uploadedFile.data;
 }
